Drop question difficulty, Rename services - delete unused file
