+++ src/models/QuestionBank/IQuestionBank.ts
@@ -8,5 +8,5 @@ export interface IQuestionBank {
   id?: string
   title?: string,
   subject?: string,
-  questionCount?: QuestionCount
+  questionCount?: number
 }
